Language button improved
