+++ src/components/languages/Languages.types.ts
@@ -1,0 +1,5 @@
+export interface Lang {
+  iconCode: "gb" | "pl" | "de";
+  code: "en" | "pl" | "de";
+  name: "English" | "Polski" | "Deutch";
+}
