+++ src/components/languages/LanguagesList.ts
@@ -1,0 +1,7 @@
+import { Lang } from "./Languages.types";
+
+export const languages: Lang[] = [
+  { code: "pl", iconCode: "pl", name: "Polski" },
+  { code: "en", iconCode: "gb", name: "English" },
+  { code: "de", iconCode: "de", name: "Deutch" },
+];
