--- src/components/languages/Types.ts
@@ -1,7 +1,0 @@
-export type SelectedLanguage = "PL" | "US" | "DE";
-export type Name = "Polski" | "English" | "Germany";
-
-export interface Lang {
-  code: SelectedLanguage;
-  name: Name;
-}
